// src/helux/src-v2-with-single-scope/typing.d.ts
export type PrimitiveItem = number | string;

export type PrimitiveSymItem = PrimitiveItem | symbol;

export type Dict<T extends any = any> = Record<PrimitiveSymItem, T>;

export type DictN<T extends any = any> = Record<number, T>;

export type Fn<T extends any = any> = (...args: any[]) => T;
// ...
export type FnType = 'watch' | 'computed';

export type ScopeType = 'static' | 'hook';

export type AsyncType = 'source' | 'task' | 'normal';

export type ReanderStatus = '1' | '2';

export type KeyUse = 'sync' | 'async' | 'both';

export type MountStatus = 1 | 2 | 3;

export interface IFnCtx {
  /** 
   * 计算函数本体，即透传给 derive 的回调函数
   */
  fn: Fn;
  /**
   * 函数唯一标记 key
   */
  fnKey: string;
  /**
   *  deriveAsync/useDeriveAsync 传入的第一个回调函数
   */
  sourceFn: Fn;
  /**
   * default: true，是否是处于第一层的函数，使用了其他计算结果时就会表标记为 false
   */
  isFirstLevel: boolean;
  isComputing: boolean;
  remainRunCount: number;
  careComputeStatus: boolean;
  /** default: false ，是否对计算结果开启记录读依赖功能，此功能仅针对 hook 里使用 useComputed 有效 */
  enableRecordResultDep: boolean;
  /**
   * 下游函数列表，即其他依赖此函数的函数列表，通常直接返回此函数结果的函数（相当于中转返回结果）会被记录到这里
   */
  downstreamFnKeys: string[];
  /** 上游函数列表，即此函数依赖的其他函数列表 */
  upstreamFnKeys: string[];
  /** 未挂载 已挂载 已卸载 */
  mountStatus: MountStatus;
  depKeys: string[];
  /** 计算函数返回的原始结果 */
  result: Dict;
  /** work for hook computed fnCtx */
  proxyResult: Dict;
  fnType: FnType;
  scopeType: ScopeType;
  /** work for hook computed fnCtx */
  updater: Fn;
  /** work for hook computed fnCtx */
  isResultReaded: boolean;
  /** 只要结果曾经读取过就记录为 true */
  isResultReadedOnce: boolean;
  /**
   * 是否返回了上游的计算结算，方便为计算结果中转机制服务
   * work for computed result transfer mechanism
   */
  returnUpstreamResult: boolean;
  /** work for hook computed fnCtx */
  renderStatus: ReanderStatus;
  /** fn ctx created timestamp */
  createTime: number;
  /** work for hook computed fnCtx  */
  shouldReplaceResult: boolean;
  /** 是否是异步的计算函数 */
  /**
   * 是否是异步的计算函数，使用了异步计算结果、返回了异步计算结果、返回了 asyncTask，满足任意一种情况都会标记为 true
   */
  isAsync: boolean;
  /** 是否是第一个异步函数，此值在isAsync=true时才有意义，使用了其他异步计算结果的异步函数，isFirstAsync 被标记为 false */
  isFirstAsync: boolean;
  /** 是否是一个中转结果的异步函数，内部用的标记 */
  isAsyncTransfer: boolean;
  transferFnCtx?: IFnCtx;
  asyncType: AsyncType,
}